requirements/test-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements/test-requirements.in
#
certifi==2023.7.22
    # via
    #   -c requirements/requirements.txt
    #   requests
charset-normalizer==3.3.2
    # via
    #   -c requirements/requirements.txt
    #   requests
coverage==7.4.4
    # via
    #   -r requirements/test-requirements.in
    #   django-coverage-plugin
django-coverage-plugin==3.1.0
    # via -r requirements/test-requirements.in
django-test-migrations==1.3.0
    # via -r requirements/test-requirements.in
exceptiongroup==1.2.0
    # via pytest
factory-boy==3.3.0
    # via -r requirements/test-requirements.in
faker==23.2.1
    # via factory-boy
idna==3.6
    # via
    #   -c requirements/requirements.txt
    #   requests
iniconfig==2.0.0
    # via pytest
packaging==23.2
    # via
    #   -c requirements/requirements.txt
    #   pytest
    #   pytest-sugar
pluggy==1.4.0
    # via pytest
pytest==8.1.1
    # via
    #   -r requirements/test-requirements.in
    #   pytest-django
    #   pytest-sugar
pytest-django==4.8.0
    # via -r requirements/test-requirements.in
pytest-sugar==1.0.0
    # via -r requirements/test-requirements.in
python-dateutil==2.8.2
    # via faker
pyyaml==6.0.1
    # via responses
requests==2.31.0
    # via
    #   -c requirements/requirements.txt
    #   responses
responses==0.25.0
    # via -r requirements/test-requirements.in
six==1.16.0
    # via
    #   -c requirements/requirements.txt
    #   python-dateutil
termcolor==2.4.0
    # via pytest-sugar
tomli==2.0.1
    # via
    #   -c requirements/requirements.txt
    #   pytest
types-requests==2.31.0.20240406
    # via -r requirements/test-requirements.in
typing-extensions==4.8.0
    # via
    #   -c requirements/requirements.txt
    #   django-test-migrations
    #   faker
urllib3==2.1.0
    # via
    #   -c requirements/requirements.txt
    #   requests
    #   responses
    #   types-requests
